Clean up utility location
