+++ serde_arrow/src/internal/serialization/date_builder.rs
@@ -5,7 +5,7 @@ use marrow::array::{Array, PrimitiveArray};
 
 use crate::internal::{
     error::{fail, set_default, try_, Context, ContextSupport, Result},
-    utils::array_ext::{new_primitive_array, ArrayExt, ScalarArrayExt},
+    utils::array_ext::{ArrayExt, ScalarArrayExt},
 };
 
 use super::{array_builder::ArrayBuilder, simple_serializer::SimpleSerializer};
@@ -53,7 +53,7 @@ impl<I: DatePrimitive> DateBuilder<I> {
     pub fn new(path: String, is_nullable: bool) -> Self {
         Self {
             path,
-            array: new_primitive_array(is_nullable),
+            array: PrimitiveArray::new(is_nullable),
         }
     }
 
@@ -65,7 +65,7 @@ impl<I: DatePrimitive> DateBuilder<I> {
     }
 
     pub fn is_nullable(&self) -> bool {
-        self.array.validity.is_some()
+        self.array.is_nullable()
     }
 
     pub fn into_array(self) -> Result<Array> {
