+++ serde_arrow/src/internal/serialization/int_builder.rs
@@ -5,7 +5,7 @@ use marrow::array::{Array, PrimitiveArray};
 use crate::internal::{
     error::{set_default, try_, Context, ContextSupport, Error, Result},
     utils::{
-        array_ext::{new_primitive_array, ArrayExt, ScalarArrayExt},
+        array_ext::{ArrayExt, ScalarArrayExt},
         NamedType,
     },
 };
@@ -22,7 +22,7 @@ impl<I: Default + 'static> IntBuilder<I> {
     pub fn new(path: String, is_nullable: bool) -> Self {
         Self {
             path,
-            array: new_primitive_array(is_nullable),
+            array: PrimitiveArray::new(is_nullable),
         }
     }
 
@@ -34,7 +34,7 @@ impl<I: Default + 'static> IntBuilder<I> {
     }
 
     pub fn is_nullable(&self) -> bool {
-        self.array.validity.is_some()
+        self.array.is_nullable()
     }
 }
