+++ serde_arrow/src/internal/serialization/timestamp_builder.rs
@@ -7,7 +7,7 @@ use marrow::{
 
 use crate::internal::{
     error::{fail, set_default, try_, Context, ContextSupport, Result},
-    utils::array_ext::{new_primitive_array, ArrayExt, ScalarArrayExt},
+    utils::array_ext::{ArrayExt, ScalarArrayExt},
 };
 
 use super::{array_builder::ArrayBuilder, simple_serializer::SimpleSerializer};
@@ -33,7 +33,7 @@ impl TimestampBuilder {
             path,
             unit,
             timezone,
-            array: new_primitive_array(is_nullable),
+            array: PrimitiveArray::new(is_nullable),
         })
     }
 
@@ -48,7 +48,7 @@ impl TimestampBuilder {
     }
 
     pub fn is_nullable(&self) -> bool {
-        self.array.validity.is_some()
+        self.array.is_nullable()
     }
 
     pub fn into_array(self) -> Result<Array> {
